+++ src/p_saveg.h
@@ -31,7 +31,7 @@ mobj_t *P_FindNewPosition(UINT32 oldposition);
 typedef struct
 {
 	UINT8 skincolor;
-	UINT8 skin;
+	UINT16 skin;
 	UINT8 botskin;
 	UINT8 botcolor;
 	INT32 score;
